Unwatching tests
